requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
ultralytics
opencv-python-headless
numpy